requirements.txt
# Core dependencies
streamlit>=1.28.0
mysql-connector-python>=8.0.0
numpy>=1.21.0
pandas>=1.5.0

# AI/ML dependencies
sentence-transformers>=2.2.0
langgraph>=0.1.0
transformers>=4.21.0
torch>=1.12.0

# LLM providers (optional - install based on your needs)
openai>=1.0.0  # For OpenAI GPT models
google-generativeai>=0.3.0  # For Google Gemini
anthropic>=0.7.0  # For Anthropic Claude

# Utilities
deep-translator>=1.11.0
python-dotenv>=1.0.0

# Development dependencies (optional)
pytest>=7.0.0
black>=23.0.0
flake8>=6.0.0
mypy>=1.0.0
